debug: isolate wait-up.js
